fix: end stream instead of destroying it
If the `source` of a sink throws an error, that error should be thrown
instead of switching to stream/event emitter style error handling.

Here we catch the thrown error, destroy the stream and clean up and listeners
then re-throw the error thrown by the source.

We don't pass the error to the stream's destroy method as that would
cause an error event to be emitted on the stream.